+++ iop/cdvd/cdfs/src/cdfs_iop.c
@@ -973,7 +973,7 @@ int cdfs_getDir(const char *pathname, const char *extensions, enum CDFS_getMode 
 
                 dir_entry++;
 
-                if (matched_entries >= req_entries)  // if we've filled the requested buffer
+                if ((unsigned int)matched_entries >= req_entries)  // if we've filled the requested buffer
                     return (matched_entries);        // then just return
 
             }  // end of the current cache block
@@ -1057,7 +1057,7 @@ int cdfs_getDir(const char *pathname, const char *extensions, enum CDFS_getMode 
 
                 dir_entry++;
 
-                if (matched_entries >= req_entries)  // if we've filled the requested buffer
+                if ((unsigned int)matched_entries >= req_entries)  // if we've filled the requested buffer
                     return (matched_entries);        // then just return
 
             }  // end of the current cache block
