+++ iop/dev9/hdpro_atad/src/hdproatad.c
@@ -897,7 +897,7 @@ int ata_io_finish(void)
             goto finish;
 
         /* Transfer each PIO data block.  */
-        while (--cmd_state->blkcount != -1) {
+        while ((int)(--cmd_state->blkcount) != -1) {
             if ((res = ata_pio_transfer(cmd_state)) < 0)
                 goto finish;
             if ((res = ata_wait_busy()) < 0)
