+++ iop/iLink/IEEE1394_disk/src/scsi.c
@@ -132,7 +132,7 @@ static int scsiInquiry(struct SBP2Device *dev, void *buffer, int size)
     } else {
         unsigned int i;
 
-        for (i = 0; i < size / 4; i++)
+        for (i = 0; i < (unsigned int)(size / 4); i++)
             ((unsigned int *)buffer)[i] = BSWAP32(((unsigned int *)buffer)[i]);
     }
